+++ lib/objectmancy.rb
@@ -5,6 +5,7 @@ require 'objectmancy/hashable'
 # Also serves as a mixin to include Objectable
 module Objectmancy
   def self.included(base)
+    base.include(Hashable)
     base.include(Objectable)
   end
 end